fix(mapping): combine workflow state and content item mapping files
